# cedar-terminology-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "CEDAR Terminology Server API",
    "title" : "CEDAR Terminology Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/bioportal/ontologies/{ontology}/properties" : {
      "get" : {
        "description" : "Get all properties from a specific ontology.",
        "operationId" : "findAllPropertiesForOntology",
        "parameters" : [ {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OntologyProperty"
                  }
                }
              }
            },
            "description" : "Every property in the ontology"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get properties",
        "tags" : [ "Properties" ]
      }
    },
...
    "/bioportal/ontologies/{ontology}/properties/{id}" : {
      "get" : {
        "description" : "Find property by id.",
        "operationId" : "findProperty",
        "parameters" : [ {
          "description" : "Property identifier. Examples: http://id.loc.gov/ontologies/bibframe/place.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OntologyProperty"
                }
              }
            },
            "description" : "The property"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find property",
        "tags" : [ "Properties" ]
      }
    },
    "/bioportal/ontologies/{ontology}/properties/{id}/children" : {
      "get" : {
        "description" : "Get property children (only for regular classes).",
        "operationId" : "findPropertyChildren",
        "parameters" : [ {
          "description" : "Property identifier. Examples: http://id.loc.gov/ontologies/bibframe/place.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OntologyProperty"
                  }
                }
              }
            },
            "description" : "The property's children"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get property children",
        "tags" : [ "Properties" ]
      }
    },
    "/bioportal/ontologies/{ontology}/properties/{id}/descendants" : {
      "get" : {
        "description" : "Get property descendants.",
        "operationId" : "findPropertyDescendants",
        "parameters" : [ {
          "description" : "Property identifier. Examples: http://id.loc.gov/ontologies/bibframe/place.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OntologyProperty"
                  }
                }
              }
            },
            "description" : "The property's descendants"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get property descendants",
        "tags" : [ "Properties" ]
      }
    },
    "/bioportal/ontologies/{ontology}/properties/{id}/parents" : {
      "get" : {
        "description" : "Get property parents.",
        "operationId" : "findPropertyParents",
        "parameters" : [ {
          "description" : "Property identifier. Examples: http://id.loc.gov/ontologies/bibframe/place.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OntologyProperty"
                  }
                }
              }
            },
            "description" : "The property's parents"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get property parents",
        "tags" : [ "Properties" ]
      }
    },
    "/bioportal/ontologies/{ontology}/properties/{id}/tree" : {
      "get" : {
        "description" : "Get property tree.",
        "operationId" : "findPropertyTree",
        "parameters" : [ {
          "description" : "Property identifier. Examples: http://id.loc.gov/ontologies/bibframe/place.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TreeNode"
                  }
                }
              }
            },
            "description" : "The paths from the ontology's roots down to the property"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get property tree",
        "tags" : [ "Properties" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "OntologyProperty" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "definitions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "hasChildren" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "labels" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "ontology" : {
            "type" : "string"
          },
          "prefLabel" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "TreeNode" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "children" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TreeNode"
            }
          },
          "hasChildren" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "obsolete" : {
            "type" : "boolean"
          },
          "ontology" : {
            "type" : "string"
          },
          "prefLabel" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}